* Fix a bunch of compiler warnings. * Remove deprecated gtk_type_class. * Fix icon update hook in launcher, this is bothering me for a long time. * Install settings desktop file in $(datadir)/xfce4/settings-dialogs too. * Fix german typo: Bug 4066. * Fix dnd to launcher: Bug 3958.
(Old svn revision: 27412)

+++ plugins/clock/clock-dialog.c
@@ -139,7 +139,7 @@ xfce_clock_dialog_tooltip_format_changed (GtkComboBox *combo,
     entry = g_object_get_data (G_OBJECT (combo), I_("entry"));
 
     /* set one of the default formats */
-    if (index < G_N_ELEMENTS (tooltip_formats))
+    if (index < (gint) G_N_ELEMENTS (tooltip_formats))
     {
         /* hide the entry */
         gtk_widget_hide (entry);
@@ -260,7 +260,7 @@ xfce_clock_dialog_digital_format_changed (GtkComboBox *combo,
     entry = g_object_get_data (G_OBJECT (combo), I_("entry"));
 
     /* set one of the default formats */
-    if (index < G_N_ELEMENTS (digital_formats))
+    if (index < (gint) G_N_ELEMENTS (digital_formats))
     {
         /* hide the entry */
         gtk_widget_hide (entry);
